requirements.txt
streamlit
pandas
numpy
torch
transformers
sentence-transformers
umap-learn
hdbscan
plotly